requirements.txt
psutil
jarbas_hive_mind>=0.10.3
pyee
requests
PyAudio==0.2.11
ovos_utils>=0.0.11
requests_futures
json_database>=0.1.3
phoneme_guesser>=0.1.0
ovos_utils>=0.0.12a9
ovos-plugin-manager>=0.0.1

ovos-ww-plugin-pocketsphinx>=0.1.0
ovos-tts-plugin-responsivevoice>=0.1.1
ovos-stt-plugin-chromium>=0.1.1
